ap_common: use intptr_t for pointer casts
this allows building on systems with pointers that aren't 16 bits long

+++ libraries/AP_Common/c++.cpp
@@ -78,10 +78,10 @@ int freeMemory()
 {
     int free_memory;
 
-    if ((int)__brkval == 0)
-        free_memory = ((int)&free_memory) - ((int)&__bss_end);
+    if ((intptr_t)__brkval == 0)
+        free_memory = ((intptr_t)&free_memory) - ((intptr_t)&__bss_end);
     else
-        free_memory = ((int)&free_memory) - ((int)__brkval);
+        free_memory = ((intptr_t)&free_memory) - ((intptr_t)__brkval);
 
     return free_memory;
 }
